♻️ (api): Refactor events folder
