+++ apps/api/src/configs/types/config-event.types.ts
@@ -1,4 +1,4 @@
-import { ConfigSchema } from "src/configs/schemas/config.schema";
+import { ConfigSchema } from "@banalize/types";
 
 export class ConfigCreatedEvent {
   constructor(public readonly config: ConfigSchema) {}
